Refactor Slider class and add MainSlider which inherits from the main slider class
