+++ src/js/main.js
@@ -1,8 +1,11 @@
 import VideoPlayer from './modules/playVideo'
-import Slider from './modules/slider'
+import MainSlider from './modules/slider/slider-main'
 
 window.addEventListener('DOMContentLoaded', () => {
-  const slider = new Slider('.page', '.next')
+  const slider = new MainSlider({
+    page: '.page',
+    buttons: '.next'
+  })
   slider.render()
   const video = new VideoPlayer('.showup .play', '.overlay')
   video.init()
