+++ src/js/modules/playVideo.js
@@ -34,7 +34,7 @@ class VideoPlayer {
       videoId: `${url}`
     });
 
-    console.log(this.player);
+    // console.log(this.player);
     this.overlay.style.display = 'flex'
   }
