+++ src/js/modules/slider/slider.js
@@ -1,0 +1,12 @@
+class Slider {
+
+  constructor({page = '', buttons = '', next = '', prev= ''} = {}) {
+    this.page = document.querySelector(page)
+    this.slides = this.page.children
+    this.buttons = document.querySelectorAll(buttons)
+    this.slideIndex = 1
+  }
+}
+
+
+export default Slider
